+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - bug with savings money
+- bug with static budget
 
 ## [3.6.0] - 2022-01-14
 ### Added